Big advance
OOP was implemented. We can generate whatever particles we want. Change easily size of box and steps each particle covers. Code is better comented. But we have a problem, particles do not crash with the wall, they go out and then come back. We need that they do not go out of the box. Maybe returning many values from the bounce functions.